% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerKerasFF.R
\name{build_shaped_mlp}
\alias{build_shaped_mlp}
\title{Shaped MLP as used in Zimmer et al. Auto Pytorch Tabular (2020)
and proposed by https://mikkokotila.github.io/slate.}
\usage{
build_shaped_mlp(task, pars)
}
\arguments{
\item{pars}{\code{\link{list}}\cr
\itemize{
\item \code{n_layers} : Number of layers.
\item \code{n_max}    : Maximum number of neurons.
}}
}
\description{
Implements 'Search Space 1' from Zimmer et al. Auto Pytorch Tabular (2020)
(https://arxiv.org/abs/2006.13799)
}
