% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerators.R
\name{combine_generators}
\alias{combine_generators}
\title{Make a DataGenerator that merges multiple DataGenerators into one.}
\usage{
combine_generators(gen1, gen2, gen1_y = TRUE, gen2_y = FALSE)
}
\arguments{
\item{gen1}{\code{DataGenerator} \cr
First Data Generator to  loop over.}

\item{gen2}{\code{DataGenerator} \cr
Second Data Generator to  loop over.}

\item{gen1_y}{\code{\link{logical}} \cr
Should target variable of \code{gen1} be returned? Default \code{TRUE}.}

\item{gen2_y}{\code{\link{logical}} \cr
Should target variable of \code{gen2} be returned? Default \code{FALSE}.}
}
\description{
Creates a Python Class that internally iterates over the data.
Generators can e.g. come from \code{make_generator_from_dataframe} or
from \code{keras::flow_images_from_dataframe}.
Need to have equal batch size, and optionally equal seeds if shuffling is used.

Returns batches of the following form:
list(
list(X1_batch, X2_batch),
Y_batch
)
or where \code{Y_batch} is Y1_batch, Y2_batch if both \code{gen._y} are \code{TRUE}, else just
e.g. Y1_batch (default).
}
