% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_helpers.R
\name{df_from_imagenet_dir}
\alias{df_from_imagenet_dir}
\title{Create a dataframe from a directory with the imagenet
directory structure.}
\usage{
df_from_imagenet_dir(dirs)
}
\arguments{
\item{dirs}{\code{\link{character}}\cr
List of dirs to create dataframes from.}
}
\value{
\code{\link{data.frame}}\cr
with columns "image" and "class" (the class).
}
\description{
The imagenet directory structure is laid out as follows
Top-level dirs: \code{train}, \code{test}, \code{valid} \cr
(you should provide path to those dirs as input) \cr
Mid-level dirs: \code{class1}, \code{class2}, \code{...} \cr
One directory for each class. The folders directly contain the images.
}
