% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_rate_finder.R
\name{find_lr}
\alias{find_lr}
\title{Plot learning rate}
\usage{
find_lr(
  learner,
  task,
  epochs = 5,
  lr_min = 10^-4,
  lr_max = 0.8,
  batch_size = 128L
)
}
\arguments{
\item{learner}{\code{\link{Learner}}\cr
An mlr3 \code{\link{Learner}} from mlr3keras.}

\item{task}{\code{\link{Task}}\cr
An mlr3 \code{\link{Task}}.}

\item{epochs}{\code{\link{integer}}\cr
Number of epochs to train for. Defaults to 5.}

\item{lr_min}{\code{\link{numeric}}\cr
Minimum learning rate to try. Defaults to 1e-4}

\item{lr_max}{\code{\link{numeric}}\cr
Maximum learning rate to try. Defaults to 0.8}

\item{batch_size}{\code{\link{numeric}}\cr
Batch size. Defaults to 128.}
}
\description{
Trains the model for few epochs, iteratively increasing learning rate.
Tries to provide insight with respect to the choice of a learning rate by iteratively increasing
learning rates from lr_min and lr_max, recording performance gains
(c.f.  Leslie N. Smith: Cyclical Learning Rates for Training Neural Networks 2015).

'find_lr' is also available via \code{Learner$lr_find(...)}.
}
