% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifKeras.R
\name{fixup_target_levels_prediction_classif}
\alias{fixup_target_levels_prediction_classif}
\title{Fix target levels}
\usage{
fixup_target_levels_prediction_classif(prob, task, out = "response")
}
\arguments{
\item{prob}{\code{\link{numeric}}\cr
The prediction to fix levels for.}

\item{task}{\code{\link{Task}}\cr
The \code{\link{Task}} to create prediction from.}

\item{out}{\code{\link{character}}\cr
Output type, either "response" or "prob".}
}
\value{
A \code{\link{PredictionClassif}}
}
\description{
Fix target levels
}
