% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerTabNet.R
\name{get_tf_num_features}
\alias{get_tf_num_features}
\title{\code{get_tf_num_features}: Get number (dimension) of tensorflow features for TabNet.}
\usage{
get_tf_num_features(task, pars)
}
\arguments{
\item{task}{\code{\link{Task}}\cr
A mlr3 Task}

\item{pars}{\cr
Model parameters, obtained through \code{your_model_name$param_set$get_values(tags = "train")}}
}
\value{
\cr
The number of features passed to TabNet.
}
\description{
\code{get_tf_num_features}: Get number (dimension) of tensorflow features for TabNet.
}
