% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_helpers.R
\name{imagepathdf_from_imagenet_dir}
\alias{imagepathdf_from_imagenet_dir}
\title{Create a dataframe with imagepaths from dirs}
\usage{
imagepathdf_from_imagenet_dir(dirs)
}
\arguments{
\item{dirs}{\code{\link{character}}\cr
List of dirs to create dataframes from.}
}
\value{
\code{\link{data.frame}}\cr
with columns "image" (an imagepath) and "class" (the class).
}
\description{
Create a dataframe with imagepaths from dirs
}
