% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{keras_reflections}
\alias{keras_reflections}
\title{Reflections mechanism for keras}
\format{
\link{environment}.
}
\usage{
keras_reflections
}
\description{
Reflections mechanism for keras
}
\details{
Used to store / extend available hyperparameter levels for options used throughout keras,
e.g. the available 'loss' for a given Learner.
}
\keyword{datasets}
