% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KerasArchitecture.R, R/LearnerKerasFF.R,
%   R/LearnerTabNet.R
\name{KerasArchitecture}
\alias{KerasArchitecture}
\alias{KerasArchitectureCustomModel}
\alias{KerasArchitectureFF}
\alias{KerasArchitectureTabNet}
\title{Keras Neural Network architecture base class}
\arguments{
\item{y_transform}{\code{\link{function}} \cr Function used to transform the targets for the model.
Default depends on the task type. See \code{y_transform} for more information.}

\item{param_set}{\code{\link{ParamSet}} \cr A methods / architecture's \code{ParamSet}.}

\item{build_arch_fun}{\code{\link{Task}} \cr Function that instantiates and compiles a model.}

\item{x_transform}{\code{\link{list}} \cr Function used to transform the data for the model.}

\item{name}{\code{\link{character}} Either 'x' or 'y'.}

\item{transform}{\code{\link{function}} \cr Function to set for the architecture}

\item{features}{\code{\link{data.table}} \cr Function that convert the features to a
form that can be passed on to \code{keras::fit()}.}

\item{target}{\code{\link{data.table}} \cr Function that convert the features to a
form that can be passed on to \code{keras::fit()}.}

\item{pars}{\code{\link{list}} \cr Parameter values, i.e. self$param_set$get_values().}

\item{...}{\code{\link{any}} \cr Additional args passed on to x_transform.}
}
\description{
A \code{KerasArchitecture} is a parametrized abstraction of a keras model.
It can be used to more easily and flexibly add architectures.

Initialize architecture

Obtain the model. Called by Learner during \code{train_internal}.

Setter method for 'x_transform' and 'y_transform'.

Transform the features before passing them on to \code{keras::fit()}.

Transform the target before passing it on to \code{keras::fit()}.

This is an architecture used for custom, user-supplied models.
The \code{model}, i.e a compiled keras model is supplied to the learner as a hyperparameter.
}
\section{Fields}{

\describe{
\item{\code{param_set}}{\code{\link{ParamSet}} \cr A method's \ architecture's \code{ParamSet}.}

\item{\code{build_arch_fun}}{\code{\link{function}} \cr Function that instantiates and compiles a model.}

\item{\code{transforms}}{\code{\link{list}} \cr The corresponding x- and y- transform.}
}}

\concept{KerasArchitectures}
