% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerTabNet.R
\name{LearnerClassifTabNet}
\alias{LearnerClassifTabNet}
\alias{mlr_learners_classif.tabnet}
\title{Keras TabNet Neural Network for Classification

Implementation of "TabNet" from the paper TabNet: Attentive Interpretable Tabular Learning (Sercan, Pfister, 2019).
See https://arxiv.org/abs/1908.07442 for details.}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3keras:LearnerClassifKeras]{mlr3keras::LearnerClassifKeras}.}
\description{
Most of the parameters can be obtained from the paper.
Some exceptions are documented here.
\itemize{
\item \code{output_dim}: Dimensions of the pen-ultimate layer(s).
\item \code{feature_dim}: Dimension of the intermediate feature representations.
}
}
\section{Construction}{
\preformatted{LearnerClassifTabNet$new()
mlr3::mlr_learners$get("classif.tabnet")
mlr3::lrn("classif.tabnet")
}
}

\examples{
learner = mlr3::lrn("classif.tabnet")
print(learner)

# available parameters:
learner$param_set$ids()
}
\references{
Sercan, A. and Pfister, T. (2019): TabNet. \url{https://arxiv.org/abs/1908.07442}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
