% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{make_data_generator}
\alias{make_data_generator}
\title{Create a data generator for a task}
\usage{
make_data_generator(
  task,
  training = TRUE,
  x_transform = function(x) x,
  y_transform = function(y) y,
  batch_size = 128,
  filter_ids = NULL
)
}
\arguments{
\item{task}{\code{\link{Task}}\cr
An mlr3 \code{\link{Task}}.}

\item{training}{If true, shuffles the data, includes y, and loops infinitely.
Training=FALSE is currently not used.}

\item{x_transform}{\code{\link{function}}\cr
Function used to transform data to a keras input format for features.}

\item{y_transform}{\code{\link{function}}\cr
Function used to transform data to a keras input format for the response.}

\item{batch_size}{\code{\link{numeric}}\cr
Batch size.}

\item{filter_ids}{\code{\link{integer}}\cr
Id's to filter.}

\item{...}{\code{\link{any}}\cr
Further arguments passed on to \code{x_transform} and \code{y_transform}}
}
\description{
Creates a data generator for a mlr3 task.
}
\examples{
require("keras")
tsk = mlr3::mlr_tasks$get("iris")
gen = reticulate::py_iterator(make_data_generator(tsk))
data = generator_next(gen) # Get next batch
}
