% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_embedding.R
\name{make_embedding}
\alias{make_embedding}
\title{Create the embedding for a dataset.
Creates an input for each categorical var, concatenates those,
Adds batch-norm to continuous vars etc.}
\usage{
make_embedding(task, embed_size = NULL, embed_dropout = 0)
}
\arguments{
\item{task}{\code{\link{Task}}\cr
An mlr3 \code{\link{Task}}.}

\item{embed_size}{\code{\link{numeric}}\cr
A numeric, either a single number determining the embedding size for
all categorical features, a named vector, determining the embedding
the size for each feature individually or \code{NULL}: Use a heuristic.
The heuristic is \code{round(1.6 * n_cat^0.56)} where n_cat is the number of levels.}

\item{embed_dropout}{\code{\link{numeric}}\cr
Dropout fraction for the embedding layer.}
}
\value{
A \code{list} of input tensors and \code{layer}: the concatenated embeddings.
}
\description{
Create the embedding for a dataset.
Creates an input for each categorical var, concatenates those,
Adds batch-norm to continuous vars etc.
}
\examples{
task = mlr3::mlr_tasks$get("boston_housing")
make_embedding(task)
}
