% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerators.R
\name{make_generator_from_dataframe}
\alias{make_generator_from_dataframe}
\title{Make a DataGenerator from a data.frame or data.table}
\usage{
make_generator_from_dataframe(
  dt,
  x_cols = NULL,
  y_cols,
  generator = keras::image_data_generator(),
  batch_size = 32L,
  shuffle = TRUE,
  seed = 1L,
  y_cols_to_categorical = TRUE,
  subset = NULL
)
}
\arguments{
\item{dt}{\link[=`data.frame`|`data.table`]{\code{data.frame}|\code{data.table}} \cr
Data container to iterate over.}

\item{x_cols}{\code{\link{character}} \cr
Names of features to be used. Defaults to all but y_cols.}

\item{y_cols}{\code{\link{character}} \cr
Target variable. Automatically converted to one-hot if "y_cols_to_categorical" is TRUE.}

\item{generator}{\code{\link{Python Object}} \cr
A generator as e.g. obtained from \code{keras::image_data_generator}.
Used for consistent train-test splits.}

\item{batch_size}{\code{\link{integer}} \cr
Batch size.}

\item{shuffle}{\code{\link{logical}} \cr
Should data be shuffled?}

\item{seed}{\code{\link{integer}} \cr
Set a seed for shuffling data.}

\item{y_cols_to_categorical}{\code{\link{logical}} \cr
Should target be converted to one-hot representation?}
}
\description{
Creates a Python Class that internally iterates over the data.
}
