% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerators.R
\name{make_generator_from_task}
\alias{make_generator_from_task}
\title{Make a DataGenerator from a \link[mlr3:Task]{mlr3::Task}}
\usage{
make_generator_from_task(
  task,
  generator = keras::image_data_generator(),
  batch_size = 32L,
  shuffle = TRUE,
  seed = 1L,
  y_cols_to_categorical = TRUE,
  subset = NULL
)
}
\arguments{
\item{task}{\code{\link{Task}} \cr
Data container to iterate over.}

\item{generator}{\code{\link{Python Object}} \cr
A generator as e.g. obtained from \code{keras::image_data_generator}.
Used for consistent train-test splits.}

\item{batch_size}{\code{\link{integer}} \cr
Batch size.}

\item{shuffle}{\code{\link{logical}} \cr
Should data be shuffled?}

\item{seed}{\code{\link{integer}} \cr
Set a seed for shuffling data.}

\item{y_cols_to_categorical}{\code{\link{logical}} \cr
Should target be converted to one-hot representation? Defaults to \code{TRUE}.}
}
\description{
Creates a Python Class that internally iterates over the data.
}
