% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGenerators.R
\name{make_train_valid_generators}
\alias{make_train_valid_generators}
\title{Create train / validation data generators from a task and params}
\usage{
make_train_valid_generators(
  task,
  x_transform,
  y_transform,
  validation_split = 1/3,
  batch_size = 128L
)
}
\arguments{
\item{task}{\code{\link{Task}}\cr
An mlr3 \code{\link{Task}}.}

\item{x_transform}{\code{\link{function}}\cr
Function used to transform data to a keras input format for features.}

\item{y_transform}{\code{\link{function}}\cr
Function used to transform data to a keras input format for the response.}

\item{validation_split}{\code{\link{numeric}}\cr
Fraction of data to use for validation.}

\item{batch_size}{\code{\link{integer}}\cr
Batch_size for the generators.}
}
\description{
Create train / validation data generators from a task and params
}
