% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3keras-package}
\alias{mlr3keras}
\alias{mlr3keras-package}
\title{mlr3keras: mlr3 keras extension}
\description{
More learners are available in the \code{mlr3learners} repository on Github (\url{https://github.com/mlr3learners}).
There also is a wiki page listing all currently available custom learners (\url{https://github.com/mlr-org/mlr3learners/wiki/Extra-Learners}).
A guide on how to create custom learners is covered in the book: \url{https://mlr3book.mlr-org.com}.
Feel invited to contribute a missing learner to the \CRANpkg{mlr3} ecosystem!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3keras}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3keras/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (0000-0001-0000-0000)
}

}
