% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_embedding.R
\name{reshape_data_embedding}
\alias{reshape_data_embedding}
\title{Reshape data for use with entity embeddings.}
\usage{
reshape_data_embedding(data)
}
\arguments{
\item{data}{\code{\link{data.table}}\cr
data.table containing the features (without target variable).}
}
\description{
Reshape data for use with entity embeddings.
}
\seealso{
reshape_task_embedding
}
