% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_embedding.R
\name{reshape_task_embedding}
\alias{reshape_task_embedding}
\title{Reshape a Task for use with entity embeddings.}
\usage{
reshape_task_embedding(task)
}
\arguments{
\item{task}{\code{\link{Task}}\cr
A mlr3 \code{\link{Task}}.}
}
\value{
A \code{list} with slots \code{data}:the reshaped data and \code{fct_levels}: the levels corresponding to each factor feature.
}
\description{
\itemize{
\item \code{logical} variables are treated as integers and converted to
either 0 or 1.
\item \code{continuous} variables are stored in a matrix "continuous"
\item \code{categorical} variables are integer encoded and stored
as a single list element each.
}
}
\examples{
task = mlr3::mlr_tasks$get("boston_housing")
reshape_task_embedding(task)
}
\seealso{
Other reshape_task_embedding: 
\code{\link{reshape_data_embedding}()}
}
\concept{reshape_task_embedding}
