% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seeds.R
\name{set_seeds}
\alias{set_seeds}
\alias{configure_session}
\title{Set Seed for \code{mlrkeras}}
\usage{
set_seeds(
  seed = 1L,
  r_s = TRUE,
  random_s = TRUE,
  python_s = TRUE,
  disable_gpu = FALSE,
  disable_parallel_cpu = FALSE
)

configure_session(disable_gpu, disable_parallel_cpu)
}
\arguments{
\item{seed}{\code{\link{integer}}\cr
A seed to be set on different platforms}

\item{r_s}{\code{\link{logical}}\cr
Should seed in R be set}

\item{random_s}{\code{\link{logical}}\cr
Should seed in random be set}

\item{python_s}{\code{\link{logical}}\cr
Should seed in python/NumPy be set}

\item{disable_gpu}{\code{\link{logical}}\cr
Should CPU parallelism be disabled}
}
\description{
\code{set_seeds}: sets a seed in Random, Python, NumPy and Tensorflow.
Futhermore it disables hash seeds, and can disable GPU and CPU parallesim.
}
