% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_callbacks.R
\name{cb_es}
\alias{cb_es}
\alias{cb_lr_scheduler}
\alias{cb_tensorboard}
\alias{cb_lr_log}
\alias{LogMetrics}
\alias{SetLogLR}
\title{\code{cb_es}: Early stopping callback}
\usage{
cb_es(monitor = "val_loss", patience = 3L)

cb_lr_scheduler()

cb_tensorboard()

cb_lr_log()
}
\arguments{
\item{monitor}{\code{\link{character}}\cr
Quantity to be monitored.}

\item{patience}{\code{\link{integer}}\cr
Number of iterations without improvement to wait before stopping.}
}
\description{
\code{cb_es}: Early stopping callback

\code{cb_lrs}: Learning rate scheduler callback

\code{cb_tb}: Tensorboard callback

\code{cb_lr_log}: Learning rate logger callback

\code{LogMetrics}: Batch-wise Metrics Logger Callback
Note, that the specified metric must be additionally supplied to
'keras::compile' or as a hyperparameter of kerasff in
order to be trackable during training.

\code{SetLogLR}: Batch-wise Metrics Setter Callback
}
