% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_kerasff.R
\docType{data}
\name{LearnerClassifkerasff}
\alias{LearnerClassifkerasff}
\alias{mlr_learners_classif.kerasff}
\title{Keras Feed Forward Neural Network}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Feed Forward Neural Network using Keras and Tensorflow.
Calls \link[keras:fit]{keras::fit} from package \CRANpkg{keras}.
}
\section{Construction}{
\preformatted{LearnerClassifkerasff$new()
mlr3::mlr_learners$get("classif.kerasff")
mlr3::lrn("classif.kerasff")
}
}

\keyword{datasets}
