% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifKerasFF.R
\docType{data}
\name{LearnerClassifKerasFF}
\alias{LearnerClassifKerasFF}
\alias{mlr_learners_classif.kerasff}
\title{Keras Feed Forward Neural Network}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\description{
Feed Forward Neural Network using Keras and Tensorflow.
Calls \link[keras:fit]{keras::fit} from package \CRANpkg{keras}.
Layers are set up as follows:
\itemize{
\item The inputs are connected to a \code{layer_dropout}, applying the \code{input_dropout}.
Afterwards, each \code{layer_dense()} is followed by a \code{layer_activation}, and
depending on hyperparameters by a \code{layer_batch_normalization} and or a
\code{layer_dropout} depending on the architecture hyperparameters.
This is repeated \code{length(layer_units)} times, i.e. one
'dense->activation->batchnorm->dropout' block is appended for each \code{layer_unit}.
}

Parameters:\cr
Most of the parameters can be obtained from the \code{keras} documentation.
Some exceptions are documented here.
\itemize{
\item \code{layer_units}: An integer vector storing the number of units in each
consecutive layer. \code{layer_units = c(32L, 32L, 32L)} results in a 3 layer
network with 32 neurons in each layer.
\item \code{initializer}: An object of class \code{tensorflow.python.ops.init_ops_v2.Initializer}.
Keras initializers start with 'initializer_...'
\item \code{optimizer}: Some optimizers and their arguments can be found below.\cr
Inherits from \code{tensorflow.python.keras.optimizer_v2}.\preformatted{"sgd"     : optimizer_sgd(lr, momentum, decay = decay),
"rmsprop" : optimizer_rmsprop(lr, rho, decay = decay),
"adagrad" : optimizer_adagrad(lr, decay = decay),
"adam"    : optimizer_adam(lr, beta_1, beta_2, decay = decay),
"nadam"   : optimizer_nadam(lr, beta_1, beta_2, schedule_decay = decay)
}
\item \code{regularizer}: Inherits from \code{tensorflow.python.keras.regularizers}.
\item \code{class_weights}: needs to be a named list of class-weights
for the different classes numbered from 0 to c-1 (for c classes).\preformatted{Example:
wts = c(0.5, 1)
setNames(as.list(wts), seq_len(length(wts)) - 1)
}
\item \code{callbacks}: A list of keras callbacks.
See \code{?callbacks}.
}
}
\section{Construction}{
\preformatted{LearnerClassifKerasFF$new()
mlr3::mlr_learners$get("classif.kerasff")
mlr3::lrn("classif.kerasff")
}
}

\examples{
learner = mlr3::lrn("classif.kerasff")
print(learner)

# available parameters:
learner$param_set$ids()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}
}
\keyword{datasets}
