% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3keras-package}
\alias{mlr3keras}
\alias{mlr3keras-package}
\title{mlr3keras: mlr3 Keras extension}
\description{
A package that connects mlr3 to keras.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3keras}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3keras/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-0000-0000}{ORCID})
}

}
