% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_embedding.R
\name{reshape_data_embedding}
\alias{reshape_data_embedding}
\title{Reshape data for use with entity embeddings.}
\usage{
reshape_data_embedding(data)
}
\arguments{
\item{data}{\code{\link{data.table}}\¢r
data.table containing the features (without target variable).}
}
\description{
\itemize{
\item \code{logical} variables are treated as integers and converted to
\code{continuous}.
\item \code{continuous} variables are stored in a matrix "continuous"
\item \code{categorical} variables are integer encoded and stored
as a single list element each.
}
}
\seealso{
reshape_task_embedding
}
